requirements.txt
boto3==1.37.38
botocore==1.37.38
contourpy==1.1.1
cycler==0.12.1
fonttools==4.57.0
importlib_resources==6.4.5
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.4
packaging==25.0
pandas==2.0.3
pillow==10.4.0
py4j==0.10.9.7
pyarrow==17.0.0
pyparsing==3.1.4
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
s3transfer==0.11.5
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.5.0
tzdata==2025.2
urllib3==1.26.20
zipp==3.20.2
